+++ rc-application/src/main/java/com/transkript/reportcard/business/service/i/TermService.kt
@@ -1,4 +1,4 @@
-package com.transkript.reportcard.business.service.interf
+package com.transkript.reportcard.business.service.i
 
 import com.transkript.reportcard.api.dto.TermDto
 import com.transkript.reportcard.api.dto.response.EntityResponse
