+++ rc-application/src/main/java/com/transkript/reportcard/business/service/i/UserService.kt
@@ -1,9 +1,7 @@
-package com.transkript.reportcard.business.service.interf
+package com.transkript.reportcard.business.service.i
 
 import com.transkript.reportcard.api.dto.UserDto
-import com.transkript.reportcard.api.dto.request.UserRequest
 import com.transkript.reportcard.api.dto.response.EntityResponse
-import com.transkript.reportcard.api.dto.response.UserResponse
 import com.transkript.reportcard.data.entity.User
 
 interface UserService {
